Update mygg.js
Added the proper display of images and binary data in general.
Cleaned up some code.
Added debug parameter.
Communication from hooked browser to mygg now is with form-multipart instead of JSON, to better control data types and not needing to urlencode and base64 encode data. To handle formData better, Busboy is now required as a dependancy.